+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.13.0 - 2023-09-01
+### Added
+- `json2xlsx.py`: support for 'aggregated' JSON input files, grouped by number of threads.
+
+### Fixed
+- `gengraph.py`: adjusted units for graphs, default is Bytes/s for encryption and decryption, and TPS for all the rest.
+
 ## 3.12.0 - 2023-08-09
 ### Added
 - test case for RSA OAEP key unwrapping (in addition to already existing RSA OAEP data decryption)